requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.135.0
Flask==3.1.3
flask-cors==6.0.2
Flask-MQTT==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
matplotlib
numpy
opencv-python
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymongo==4.10.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.47
starlette==0.52.1
torch
torchvision
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
dnspython==2.7.0